requirements.txt
pandas
scikit-learn
flask
flask-socketio
matplotlib
numpy
pyspark
kafka-python
pymongo
requests